Directory Parse warning (#64)
* added warning to directory parse

added logging warning to parse_directory so that it will print out files that cannot be opened by lasio. It then moves on to the next file in the directory and parses that one.

* Update keyword_tree.py

* added empty las file and test for logging warning

added an empty las file (nothing inside the file at all) that will cause lasio read issues which in turn should call the logging warning in parse_directory. New test catches that warning